requirements.txt
sentence-transformers
langchain
flask
pypdf
python-dotenv
pinecone[grpc]
langchain_huggingface
langchain-pinecone
langchain_community
langchain_openai
langchain_experimental
-e .